user/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
Flask-Cors==3.0.10
flask-restx==0.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
jsonschema==3.2.0
pycparser==2.20
PyMySQL==1.0.2
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.7
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
uWSGI==2.0.19.1
Werkzeug==1.0.1

